Apply review suggestions.
