+++ src/main/java/com/direwolf20/buildinggadgets/common/schema/TemplateData.java
@@ -7,13 +7,18 @@ import net.minecraft.util.math.BlockPos;
 import javax.annotation.Nullable;
 import java.util.Objects;
 
-public final class MutableBlockData {
+/**
+ * This class is a mutable wrapper around {@link BlockPos position}, {@link BlockState state} and {@link CompoundNBT nbt}
+ * and therefore represents a single data point in a {@link Template}. Notice that because it is mutable this class should
+ * not be stored directly when outside code has references to it.
+ */
+public final class TemplateData {
     private BlockPos pos;
     private BlockState state;
     @Nullable
     private CompoundNBT tileNbt;
 
-    public MutableBlockData(BlockPos pos, BlockState state, @Nullable CompoundNBT tileNbt) {
+    public TemplateData(BlockPos pos, BlockState state, @Nullable CompoundNBT tileNbt) {
         setInformation(pos, state, tileNbt);
     }
 
@@ -30,27 +35,27 @@ public final class MutableBlockData {
         return tileNbt;
     }
 
-    public MutableBlockData setInformation(BlockPos pos, BlockState state, @Nullable CompoundNBT tileNbt) {
+    public TemplateData setInformation(BlockPos pos, BlockState state, @Nullable CompoundNBT tileNbt) {
         this.pos = Objects.requireNonNull(pos, "Cannot have BlockData with null position");
         this.state = Objects.requireNonNull(state, "Cannot have BlockData with null BlockState");
         this.tileNbt = tileNbt;
         return this;
     }
 
-    public MutableBlockData setInformation(BlockPos pos, BlockState state) {
+    public TemplateData setInformation(BlockPos pos, BlockState state) {
         return setInformation(pos, state, null);
     }
 
-    public BlockData toImmutable() {
-        return BlockData.create(pos, state, tileNbt);
+    public TemplateData copy() {
+        return new TemplateData(pos.toImmutable(), state, tileNbt != null ? tileNbt.copy() : null);
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
-        if (! (o instanceof MutableBlockData)) return false;
+        if (! (o instanceof TemplateData)) return false;
 
-        final MutableBlockData that = (MutableBlockData) o;
+        final TemplateData that = (TemplateData) o;
 
         if (! pos.equals(that.pos)) return false;
         if (! state.equals(that.state)) return false;
